requirements.txt
jsonschema==4.3.3
PyYAML==5.4.1
datamodel-code-generator==0.11.15
json-schema-for-humans==0.39.5
typer==0.4.0
Jinja2==3.0.3
netaddr==0.8.0
ansible==5.1.0
pytest==6.2.5
pytest-html
types-PyYAML
requests
rich==11.0.0